Add a function to determine the winner
The getComputerSelection and getPlayerSelection functions have been
modified to return a number instead of a string. This makes finding
the winner simpler. The number corresponds to the index in the hand
array. So if the user writes "rOCk", the value is 0 because
hand[0] === "rock". The program now compares indexes (plural of index?)
to determine the winner of the round.

The output message needs to be set according to the standards specified
on the project web page. Something like "You win! Scissors beat paper."
or "You lose! Paper beats rock." So far it only tell if the computer
or the human has won. Baby steps.