Add the final touch
The program can now play until a player has five points and start
a new game if the user wants. I meets all objectives, or so I believe
right now. I will check tomorrow to make sure, but it seems to me
like this is ready to ship. How many programmers said that and
regretted it almost immediately, I wonder.