Add rematch feature
When the score of one player reaches 5, the play buttons are
disabled and two new buttons show up. These new buttons offer
the player a rematch. The player can decline, in which case,
the play buttons disappear and the page must be reloaded to
play again. If the player accepts, the rematch buttons disappear
and the play buttons are enabled once more.

The next aim is to display the hand of both players on the page.
Each round's result should be kept as a record until a new game
starts. I will create a div to hold the rounds' results and push
them down as the newer results come in.